# [Performance] Improve the performance of Radius.tsx by 2x