--- a/athena/MapData.tsx
+++ b/athena/MapData.tsx
@@ -416,7 +416,8 @@ export default class MapData {
     value: T,
   ): T {
     const { map, size } = this;
-    for (let i = 0; i < map.length; i++) {
+    const len = map.length;
+    for (let i = 0; i < len; i++) {
       value = fn.call(this, value, indexToVector(i, size.width), i);
     }
     return value;
@@ -450,7 +451,8 @@ export default class MapData {
     value: T,
   ): T {
     const { map, modifiers, size } = this;
-    for (let i = 0; i < map.length; i++) {
+    const len = map.length;
+    for (let i = 0; i < len; i++) {
       const field = map[i];
       if (typeof field === 'number') {
         value = fn.call(
@@ -502,7 +504,8 @@ export default class MapData {
     value: T,
   ): T {
     const { decorators, size } = this;
-    for (let i = 0; i < decorators.length; i++) {
+    const len = decorators.length;
+    for (let i = 0; i < len; i++) {
       const decorator = getDecorator(decorators[i]);
       if (decorator) {
         value = fn.call(