--- a/athena/MapData.tsx
+++ b/athena/MapData.tsx
@@ -173,13 +173,13 @@ const toPlayer = (object: AnyEntity): PlayerID =>
       : object.id;
 
 export default class MapData {
-  protected id: string;
   private players: Map<PlayerID, Player>;
   private playerToTeam: Map<PlayerID, number>;
   private _firstPlayer: PlayerID = 0;
   private _hasNeutralUnits: boolean | null = null;
   private _activeUnitTypes: ReadonlyMap<PlayerID, ActiveUnitTypes> | null =
     null;
+  protected unitAccessibilityCache: Map<string | number, boolean>;
 
   constructor(
     public readonly map: TileMap,
@@ -194,7 +194,6 @@ export default class MapData {
     public readonly buildings: ImmutableMap<Vector, Building>,
     public readonly units: ImmutableMap<Vector, Unit>,
   ) {
-    this.id = JSON.stringify(this.buildings.toJSON());
     this.players = new Map(
       teams.flatMap((team) => team.players).sortBy(({ id }) => id),
     );
@@ -208,10 +207,11 @@ export default class MapData {
       }),
     );
     this.playerToTeam.set(0, -1);
+    this.unitAccessibilityCache = new Map();
   }
 
-  getId() {
-    return this.id;
+  getCache() {
+    return this.unitAccessibilityCache;
   }
 
   contains(size: { x: number; y: number }) {