--- a/athena/Radius.tsx
+++ b/athena/Radius.tsx
@@ -1,5 +1,4 @@
 import FastPriorityQueue from 'fastpriorityqueue';
-import LRUCache from 'mnemonist/lru-cache';
 import { Skill } from './info/Skill.tsx';
 import { TileInfo, TileTypes } from './info/Tile.tsx';
 import { UnitInfo } from './info/Unit.tsx';
@@ -38,18 +37,7 @@ export const RadiusItem = (
   vector,
 });
 
-let lruCacheMap: LRUCache<string | number, boolean>;
-let currentMapId: string;
-
-function getCache(map: MapData) {
-  // Reinstantiate the cache if the map changes
-  if (currentMapId !== map.getId()) {
-    const cacheCapacity = (map.size.width * map.size.height) / 2; // No need to cache the entire map
-    lruCacheMap = new LRUCache(cacheCapacity);
-    currentMapId = map.getId();
-  }
-  return lruCacheMap;
-}
+let currentUnitId: number;
 
 function getCostBase(map: MapData, unit: Unit, vector: Vector, index?: number) {
   const tileInfo = map.maybeGetTileInfo(vector, undefined, index);
@@ -77,12 +65,13 @@ function isAccessibleBase(map: MapData, unit: Unit, vector: Vector) {
 }
 
 function isAccessible(map: MapData, unit: Unit, vector: Vector) {
-  const cache = getCache(map);
+  const cache = map.getCache();
   const key = vector.hashCode();
   let accessible = cache.get(key);
-  if (accessible != null) {
+  if (accessible !== undefined && currentUnitId === unit.id) {
     return accessible;
   }
+  currentUnitId = unit.id;
   accessible = isAccessibleBase(map, unit, vector);
   cache.set(key, accessible);
   return accessible;