--- a/athena/package.json
+++ b/athena/package.json
@@ -14,7 +14,6 @@
     "@nkzw/immutable-map": "^1.2.2",
     "array-shuffle": "^3.0.0",
     "fastpriorityqueue": "^0.7.5",
-    "mnemonist": "^0.39.8",
     "skmeans": "^0.11.3"
   },
   "devDependencies": {