Closes #9

I used the [`moveable`](https://github.com/nkzw-tech/athena-crisis/blob/f990b25b5cf5340b45a6b50308c2e87acb099946/athena/Radius.tsx#L148) function as a benchmark entry point so other results may vary.
Consistently achieved just a bit over a 2x improvement with few to no core changes.

_Platform_
Windows_NT 10.0.19045 x64
Node.JS: 20.12.2
V8: 11.3.244.8-node.19

| fast moveable    | 0% (base)     | (16.479 ops/sec) | (avg: 60μs)  | 
|------------------|--------|------------------|--------------|
| regular moveable | -56,9% | (7102 ops/sec)   | (avg: 140μs) |

![img](https://image-charts.com/chart.js/2.8.0?bkg=white&c=%7B%22type%22%3A%22bar%22%2C%22data%22%3A%7B%22labels%22%3A%5B%22fast%20moveable%22%2C%22regular%20moveable%22%5D%2C%22datasets%22%3A%5B%7B%22label%22%3A%22Dataset%201%22%2C%22backgroundColor%22%3A%22rgba%2854%2C%20162%2C%20235%2C%200.5%29%22%2C%22borderColor%22%3A%22rgb%2854%2C%20162%2C%20235%29%22%2C%22borderWidth%22%3A1%2C%22data%22%3A%5B16479.287532356466%2C7102.202061644472%5D%7D%5D%7D%2C%22options%22%3A%7B%22responsive%22%3Afalse%2C%22legend%22%3A%7B%22display%22%3Afalse%2C%22position%22%3A%22top%22%7D%2C%22title%22%3A%7B%22display%22%3Atrue%2C%22text%22%3A%22moveable%7C%28ops%2Fsec%29%22%7D%2C%22layout%22%3A%7B%22padding%22%3A20%7D%7D%7D)