@cpojer 2024-07-08
That's awesome, thank you for the PR! I left a few comments to make sure we aren't breaking something across maps by caching data globally. Could you also share the benchmark you used?