@rortan134 2024-07-08
> Could you also share the benchmark you used?

I used the [benchmarkify](https://www.npmjs.com/package/benchmarkify) library with some mock data from the tests ([Radius.test.tsx](https://github.com/nkzw-tech/athena-crisis/blob/main/athena/__tests__/Radius.test.tsx))