@cpojer 2024-07-08
> > Could you also share the benchmark you used?
> 
> I used the [benchmarkify](https://www.npmjs.com/package/benchmarkify) library with the mock data from the first test in ([Radius.test.tsx](https://github.com/nkzw-tech/athena-crisis/blob/main/athena/__tests__/Radius.test.tsx?rgh-link-date=2024-07-08T02%3A24%3A43Z))

That's great. Can you also share the code so we can repro it?