@rortan134 2024-07-08
> That's great. Can you also share the code so we can repro it?

Absolutely, https://github.com/rortan134/athena-crisis/blob/radius-perf-bmark/athena/__tests__/RadiusBenchmark.test.tsx
Just run the test script