@cpojer 2024-07-09
Thanks for the updates!

I realized that we can actually avoid all of the extra complexity and don't need to add an LRUCache. The radius cache is specific to a MapData instance, and MapData is immutable. It's just a cache to speed things up, so how about we just add the cache interface to `MapData` directly? Then it will be garbage collected automatically.

However, this led me to uncover a gap in this optimization. `calculateRadius` takes four inputs: A map, a unit, a starting Vector and the radius. All four of those can be different at any time, like for example we might want to get a Flamethrower's radius at the center of a desert map, or a Sniper's radius in a volcano map at the left top. However, we might also check different units at one position on the same map, and each unit has different movement costs on the same field, like for example an ocean might be inaccessible to a Flamethrower but a Fighter Jet can just fly over it. This means that caching on a per-map and per-field basis will not work. Let me know if you'd like me to add a bunch of test cases to verify these behaviors.