@rortan134 2024-07-09
I see.

> Let me know if you'd like me to add a bunch of test cases to verify these behaviors.

A couple test cases would be great