@cpojer 2024-07-10
Thanks for the updates, I think we are getting closer to something that can be merged. As for the benchmark, it's not super representative to run it on the same inputs – if that was a common case we could just cache the outputs of the function completely!

Here is a more representative benchmark that calculates the radius for each unit on the map:

```
test('benchmark moveable radius', async () => {
  const testMap = startMap.copy({
    units: startMap.units.delete(vec(11, 8)),
  });
  
  const benchmark = new Benchmark('benchmark moveable').printHeader();
  benchmark
    .createSuite('starting benchmark', { time: 10_000 })
    .add('moveable', () => {
      for (const [vector, unit] of testMap.units) {
        moveable(testMap, unit, vector, 7)
      }
    })
    .add('fast moveable', () => {
      for (const [vector, unit] of testMap.units) {
        fastMoveable(testMap, unit, vector, 7)
      }
    })
  await benchmark.run();
});
```

It might be worth randomizing the order at the beginning of the run, but this should be a good start.