@cpojer 2024-07-11
I added the benchmarks right to the repo using `pnpm vitest bench`: https://github.com/nkzw-tech/athena-crisis/commit/5d3bc85d50732b6846885f99c26da53655f713f3

Feel free to extend them as well.