@cpojer 2024-08-02
@rortan134 are you planning on getting back to this?